remove old texture scripts and dae models
