+++ demo/data_texture_preview.collection
@@ -16,6 +16,8 @@ embedded_instances {
   "    z: 0.0\n"
   "    w: 1.0\n"
   "  }\n"
+  "  property_decls {\n"
+  "  }\n"
   "}\n"
   "embedded_components {\n"
   "  id: \"bg\"\n"
@@ -44,12 +46,6 @@ embedded_instances {
   "  x: 128.0\\n"
   "  y: 32.0\\n"
   "  z: 0.0\\n"
-  "  w: 0.0\\n"
-  "}\\n"
-  "scale {\\n"
-  "  x: 1.0\\n"
-  "  y: 1.0\\n"
-  "  z: 1.0\\n"
   "  w: 0.0\\n"
   "}\\n"
   "color {\\n"
@@ -129,25 +125,7 @@ embedded_instances {
   "    z: 0.0\n"
   "    w: 1.0\n"
   "  }\n"
-  "  properties {\n"
-  "    id: \"segment_length_max\"\n"
-  "    value: \"6.0\"\n"
-  "    type: PROPERTY_TYPE_NUMBER\n"
-  "  }\n"
-  "  properties {\n"
-  "    id: \"segment_length_min\"\n"
-  "    value: \"6.0\"\n"
-  "    type: PROPERTY_TYPE_NUMBER\n"
-  "  }\n"
-  "  properties {\n"
-  "    id: \"fade_tail_alpha\"\n"
-  "    value: \"12.0\"\n"
-  "    type: PROPERTY_TYPE_NUMBER\n"
-  "  }\n"
-  "  properties {\n"
-  "    id: \"shrink_tail_width\"\n"
-  "    value: \"true\"\n"
-  "    type: PROPERTY_TYPE_BOOLEAN\n"
+  "  property_decls {\n"
   "  }\n"
   "}\n"
   "components {\n"
@@ -164,18 +142,16 @@ embedded_instances {
   "    z: 0.0\n"
   "    w: 1.0\n"
   "  }\n"
+  "  property_decls {\n"
+  "  }\n"
   "}\n"
   "embedded_components {\n"
   "  id: \"trail_model\"\n"
-  "  type: \"model\"\n"
-  "  data: \"mesh: \\\"/hyper_trails/models/trail_16.dae\\\"\\n"
-  "material: \\\"/hyper_trails/materials/trail.material\\\"\\n"
-  "textures: \\\"/hyper_trails/textures/data/texture0_1.png\\\"\\n"
+  "  type: \"mesh\"\n"
+  "  data: \"material: \\\"/hyper_trails/materials/trail.material\\\"\\n"
+  "vertices: \\\"/hyper_trails/models/trail_model.buffer\\\"\\n"
   "textures: \\\"/hyper_trails/textures/white.png\\\"\\n"
-  "skeleton: \\\"\\\"\\n"
-  "animations: \\\"\\\"\\n"
-  "default_animation: \\\"\\\"\\n"
-  "name: \\\"unnamed\\\"\\n"
+  "primitive_type: PRIMITIVE_TRIANGLE_STRIP\\n"
   "\"\n"
   "  position {\n"
   "    x: 0.0\n"
