+++ demo/demo_physics.collection
@@ -157,40 +157,16 @@ embedded_instances {
   "    z: 0.0\n"
   "    w: 1.0\n"
   "  }\n"
-  "  properties {\n"
-  "    id: \"trail_width\"\n"
-  "    value: \"10.0\"\n"
-  "    type: PROPERTY_TYPE_NUMBER\n"
-  "  }\n"
-  "  properties {\n"
-  "    id: \"fade_tail_alpha\"\n"
-  "    value: \"12.0\"\n"
-  "    type: PROPERTY_TYPE_NUMBER\n"
-  "  }\n"
-  "  properties {\n"
-  "    id: \"shrink_tail_width\"\n"
-  "    value: \"true\"\n"
-  "    type: PROPERTY_TYPE_BOOLEAN\n"
-  "  }\n"
-  "  properties {\n"
-  "    id: \"auto_update\"\n"
-  "    value: \"false\"\n"
-  "    type: PROPERTY_TYPE_BOOLEAN\n"
-  "  }\n"
   "  property_decls {\n"
   "  }\n"
   "}\n"
   "embedded_components {\n"
   "  id: \"trail_model\"\n"
-  "  type: \"model\"\n"
-  "  data: \"mesh: \\\"/hyper_trails/models/trail_16.dae\\\"\\n"
-  "material: \\\"/hyper_trails/materials/trail.material\\\"\\n"
-  "textures: \\\"/hyper_trails/textures/data/texture0_3.png\\\"\\n"
+  "  type: \"mesh\"\n"
+  "  data: \"material: \\\"/hyper_trails/materials/trail.material\\\"\\n"
+  "vertices: \\\"/hyper_trails/models/trail_model.buffer\\\"\\n"
   "textures: \\\"/hyper_trails/textures/white.png\\\"\\n"
-  "skeleton: \\\"\\\"\\n"
-  "animations: \\\"\\\"\\n"
-  "default_animation: \\\"\\\"\\n"
-  "name: \\\"unnamed\\\"\\n"
+  "primitive_type: PRIMITIVE_TRIANGLE_STRIP\\n"
   "\"\n"
   "  position {\n"
   "    x: 0.0\n"
