+++ hyper_trails/hyper_trail.go
@@ -1,6 +1,6 @@
 components {
   id: "trail_maker"
-  component: "/hyper_trails/trail_maker_new.script"
+  component: "/hyper_trails/trail_maker.script"
   position {
     x: 0.0
     y: 0.0
@@ -14,42 +14,88 @@ components {
   }
   properties {
     id: "trail_width"
-    value: "24.0"
+    value: "10.0"
     type: PROPERTY_TYPE_NUMBER
   }
   properties {
-    id: "segment_length_max"
-    value: "200.0"
+    id: "fade_tail_alpha"
+    value: "12.0"
     type: PROPERTY_TYPE_NUMBER
   }
   properties {
-    id: "segment_length_min"
-    value: "200.0"
-    type: PROPERTY_TYPE_NUMBER
-  }
-  properties {
-    id: "points_count"
-    value: "4.0"
-    type: PROPERTY_TYPE_NUMBER
-  }
-  properties {
-    id: "points_limit"
-    value: "0.0"
-    type: PROPERTY_TYPE_NUMBER
-  }
-  properties {
-    id: "texture_tiling"
-    value: "false"
+    id: "shrink_tail_width"
+    value: "true"
     type: PROPERTY_TYPE_BOOLEAN
+  }
+}
+components {
+  id: "static_responce"
+  component: "/hyper_trails/static_responce.script"
+  position {
+    x: 0.0
+    y: 0.0
+    z: 0.0
+  }
+  rotation {
+    x: 0.0
+    y: 0.0
+    z: 0.0
+    w: 1.0
   }
 }
 embedded_components {
   id: "trail_model"
   type: "mesh"
-  data: "material: \"/hyper_trails/materials_new/trail.material\"\n"
+  data: "material: \"/hyper_trails/materials/trail.material\"\n"
   "vertices: \"/hyper_trails/models/trail_model.buffer\"\n"
-  "textures: \"/demo/assets/images/logo_text_texture.png\"\n"
+  "textures: \"/hyper_trails/textures/white.png\"\n"
   "primitive_type: PRIMITIVE_TRIANGLE_STRIP\n"
+  ""
+  position {
+    x: 0.0
+    y: 0.0
+    z: 0.0
+  }
+  rotation {
+    x: 0.0
+    y: 0.0
+    z: 0.0
+    w: 1.0
+  }
+}
+embedded_components {
+  id: "collisionobject"
+  type: "collisionobject"
+  data: "collision_shape: \"\"\n"
+  "type: COLLISION_OBJECT_TYPE_DYNAMIC\n"
+  "mass: 1.0\n"
+  "friction: 0.1\n"
+  "restitution: 0.5\n"
+  "group: \"default\"\n"
+  "mask: \"default\"\n"
+  "embedded_collision_shape {\n"
+  "  shapes {\n"
+  "    shape_type: TYPE_SPHERE\n"
+  "    position {\n"
+  "      x: 0.0\n"
+  "      y: 0.0\n"
+  "      z: 0.0\n"
+  "    }\n"
+  "    rotation {\n"
+  "      x: 0.0\n"
+  "      y: 0.0\n"
+  "      z: 0.0\n"
+  "      w: 1.0\n"
+  "    }\n"
+  "    index: 0\n"
+  "    count: 1\n"
+  "  }\n"
+  "  data: 3.1885\n"
+  "}\n"
+  "linear_damping: 0.0\n"
+  "angular_damping: 0.0\n"
+  "locked_rotation: false\n"
+  "bullet: false\n"
   ""
   position {
     x: 0.0
