+++ hyper_trails/materials/trail.fp
@@ -1,7 +1,6 @@
 varying lowp vec4 var_tint;
 varying mediump vec2 var_texcoord0;
 
-//uniform highp sampler2D texture0;
 uniform lowp sampler2D texture1;
 
 void main()
