--- hyper_trails/materials_new/trail.material
@@ -1,33 +1,0 @@
-name: "hyper_trails_trail"
-tags: "tile"
-vertex_program: "/hyper_trails/materials_new/trail.vp"
-fragment_program: "/hyper_trails/materials_new/trail.fp"
-vertex_space: VERTEX_SPACE_LOCAL
-vertex_constants {
-  name: "world_view_proj"
-  type: CONSTANT_TYPE_WORLDVIEWPROJ
-  value {
-    x: 0.0
-    y: 0.0
-    z: 0.0
-    w: 0.0
-  }
-}
-vertex_constants {
-  name: "uv_opts"
-  type: CONSTANT_TYPE_USER
-  value {
-    x: 1.0
-    y: 0.0
-    z: 1.0
-    w: 0.0
-  }
-}
-samplers {
-  name: "texture1"
-  wrap_u: WRAP_MODE_REPEAT
-  wrap_v: WRAP_MODE_REPEAT
-  filter_min: FILTER_MODE_MIN_LINEAR_MIPMAP_LINEAR
-  filter_mag: FILTER_MODE_MAG_LINEAR
-  max_anisotropy: 1.0
-}
