--- hyper_trails/materials_new/trail.vp
@@ -1,17 +1,0 @@
-attribute mediump vec3 position;
-attribute mediump vec2 texcoord0;
-attribute lowp vec4 tint;
-
-uniform mediump vec4 uv_opts;
-uniform mediump mat4 world_view_proj;
-
-
-varying mediump vec2 var_texcoord0;
-varying lowp vec4 var_tint;
-
-void main() {
-    var_tint = tint;
-    var_texcoord0.x = texcoord0.x;
-    var_texcoord0.y = texcoord0.y * uv_opts.x;
-    gl_Position = world_view_proj * vec4(position, 1.0);
-}
