+++ hyper_trails/static_responce.script
@@ -1,0 +1,7 @@
+function on_message(self, message_id, message, sender)
+	if message_id == hash("collision_response") then
+		timer.delay(0.1, false, function()
+			go.delete()
+		 end)
+	end
+end
