+++ hyper_trails/trail_maker.script
@@ -1,4 +1,3 @@
-
 local trail_maker = require("hyper_trails.trail_maker")
 
 go.property("use_world_position", false)
@@ -17,17 +16,15 @@ go.property("auto_update", true)
 
 local function reset_trail(self)
 	trail_maker.init_vars(self)
+	trail_maker.init_buffers(self)
 	trail_maker.init_props(self)
-	trail_maker.create_texture(self)
 	trail_maker.init_data_points(self)
 	trail_maker.draw_trail(self)
 end
 
 function init(self)
-	reset_trail(self)
-end
-
-function final(self)
+	self.is_ready = false
+	msg.post("/trails_pool#trails_pool", "get_buffer")
 end
 
 local function update_trail(self, dt)
@@ -36,6 +33,9 @@ local function update_trail(self, dt)
 end
 
 function update(self, dt)
+	if not self.is_ready then
+		return
+	end
 	self.dt = dt
 	if self.auto_update then
 		update_trail(self, dt)
@@ -46,9 +46,20 @@ local UPDATE = hash("update")
 local RESET = hash("reset")
 
 function on_message(self, message_id, message, sender)
-	if message_id == UPDATE then
+	if message_id == UPDATE and self.is_ready then
 		update_trail(self, message.dt or self.dt)
+		
 	elseif message_id == RESET then
 		reset_trail(self)
+		
+	elseif message_id == hash("buffer_index") then
+		self.mesh_vertices_resource = go.get("/trails_pool#trails_pool", "buffer_" .. message.index)
+		reset_trail(self)
+		self.buffer_index = message.index
+		self.is_ready = true
 	end
+end
+
+function final(self)
+	msg.post("/trails_pool#trails_pool", "releace_buffer", {index=self.buffer_index})
 end
